services/aggregation/requirements.txt
aio-pika==9.3.1
psycopg2-binary==2.9.9
prometheus-client==0.19.0
python-dotenv==1.0.0
